refactor: Put api services into logical grouping based on resources
